商品添加删除更新

+++ app/Http/Controllers/Api/ShopController.php
@@ -20,7 +20,7 @@ class ShopController extends Controller
     /*
     * 我的商店列表 */
     /**
-     * @SWG\Get(path="/api/shop", tags={"Shop"},summary="我的商店",description="商店列表",operationId="",produces={"application/json"},
+     * @SWG\Get(path="/api/shop", tags={"Shop 商店类Api"},summary="我的商店",description="商店列表",operationId="",produces={"application/json"},
      *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
      *   @SWG\Response(response="default", description="操作成功")
      * )
@@ -31,7 +31,7 @@ class ShopController extends Controller
     /*
      * 增加我的商店*/
     /**
-     * @SWG\Post(path="/api/shop", tags={"Shop"},summary="创建我的商店",description="创建商店",operationId="",produces={"application/json"},
+     * @SWG\Post(path="/api/shop", tags={"Shop 商店类Api"},summary="创建我的商店",description="创建商店",operationId="",produces={"application/json"},
      *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
      *   @SWG\Parameter(in="formData",name="name",type="string",description="商店名称",required=true),
      *   @SWG\Parameter(in="formData",name="logo",type="string",description="用户头像",required=true),
@@ -59,7 +59,7 @@ class ShopController extends Controller
     }
 
     /**
-     * @SWG\Put(path="/api/shop/{id}", tags={"Shop"},summary="更新商店信息",description="更新商店信息",operationId="",produces={"application/json"},
+     * @SWG\Put(path="/api/shop/{id}", tags={"Shop 商店类Api"},summary="更新商店信息",description="更新商店信息",operationId="",produces={"application/json"},
      *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
      *   @SWG\Parameter(in="header",name="Content-Type",type="string",description="Content-Type",default="application/x-www-form-urlencoded"),
      *   @SWG\Parameter(in="path",name="id",type="string",description="id",required=true),
@@ -76,6 +76,10 @@ class ShopController extends Controller
     /*
      * 更新店铺信息*/
     public function update(Request $request, $id){
+        if($this->isIllegal($request->user(),$id)){
+            return $this->sendErrorMsg('这不是您的商店！非法操作！');
+        }
+
         $prev = Shop::find($id);
         if(!$prev->editable){
             return $this->sendErrorMsg('当前不可编辑！');
@@ -85,15 +89,8 @@ class ShopController extends Controller
         if($validator->fails()){
             return $this->sendValidateErrorMsg($validator);
         }
-        $prev->name = $data['name'];
-        $prev->logo = $data['logo'];
-        $prev->location = $data['location'];
-        $prev->type = $data['type'];
-        $prev->description = $data['description'];
-        $prev->open_at = $data['open_at'];
-        $prev->close_at = $data['close_at'];
         try{
-            $prev->save();
+            $prev->update($data);
             return $this->sendSuccessMsg('保存成功', $prev);
         }catch(\Exception $e){
             return $this->sendSqlErrorMsg($e);
@@ -103,17 +100,20 @@ class ShopController extends Controller
     /*
      * 删除店铺 */
     /**
-     * @SWG\Delete(path="/api/shop/{id}", tags={"Shop"},summary="删除商店",description="删除商店",operationId="",produces={"application/json"},
+     * @SWG\Delete(path="/api/shop/{id}", tags={"Shop 商店类Api"},summary="删除商店",description="删除商店",operationId="",produces={"application/json"},
      *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
      *   @SWG\Parameter(in="path",name="id",type="string",description="id",required=true),
      *   @SWG\Response(response="default", description="操作成功")
      * )
      */
-    /*
     public function delete(Request $request, $id){
+        if($this->isIllegal($request->user(),$id)){
+            return $this->sendErrorMsg('这不是您的商店！非法操作！');
+        }
         try{
             DB::transaction(function () use($id){
                 Shop::find($id)->delete();
+                Goods::where('shop_id',$id)->delete();
                 ShopCertify::where('shop_id',$id)->delete();
             });
             return $this->sendSuccessMsg('删除成功！');
@@ -125,7 +125,7 @@ class ShopController extends Controller
     /*
      * 店铺详情*/
     /**
-     * @SWG\Get(path="/api/shop/{id}", tags={"Shop"},summary="商店详情",description="商店详情",operationId="",produces={"application/json"},
+     * @SWG\Get(path="/api/shop/{id}", tags={"Shop 商店类Api"},summary="商店详情",description="商店详情",operationId="",produces={"application/json"},
      *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
      *   @SWG\Parameter(in="path",name="id",type="string",description="id",required=true),
      *   @SWG\Response(response="default", description="操作成功")
@@ -133,14 +133,14 @@ class ShopController extends Controller
      */
     public function detail(Request $request, $id){
        $shop = Shop::find($id);
-       $shop['goods'] = $shop->goods;
+       //$shop['goods'] = $shop->goods;
        return $shop;
     }
 
     /*
    *  我的店铺审核 */
     /**
-     * @SWG\Get(path="/api/shop/certify", tags={"Shop"},summary="我的审核",description="我的审核",operationId="",produces={"application/json"},
+     * @SWG\Get(path="/api/shop/certify", tags={"Shop 商店类Api"},summary="我的审核",description="我的审核",operationId="",produces={"application/json"},
      *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
      *   @SWG\Response(response="default", description="操作成功")
      * )
@@ -158,13 +158,16 @@ class ShopController extends Controller
     /*
     *  申请营业审批 */
     /**
-     * @SWG\Get(path="/api/shop/certify/{id}", tags={"Shop"},summary="申请营业审批",description="申请营业审批",operationId="",produces={"application/json"},
+     * @SWG\Get(path="/api/shop/certify/{id}", tags={"Shop 商店类Api"},summary="申请营业审批",description="申请营业审批",operationId="",produces={"application/json"},
      *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
      *   @SWG\Parameter(in="path",name="id",type="string",description="id",required=true),
      *   @SWG\Response(response="default", description="操作成功")
      * )
      */
     public function certify(Request $request, $id){
+        if($this->isIllegal($request->user(),$id)){
+            return $this->sendErrorMsg('这不是您的商店！非法操作！');
+        }
         $shop = Shop::find($id);
         $user = $request->user();
         /* 实名认证检测 */
@@ -196,13 +199,16 @@ class ShopController extends Controller
     /*
      *  取消营业审批 */
     /**
-     * @SWG\Delete(path="/api/shop/certify/{id}", tags={"Shop"},summary="删除营业审批申请",description="删除营业审批申请",operationId="",produces={"application/json"},
+     * @SWG\Delete(path="/api/shop/certify/{id}", tags={"Shop 商店类Api"},summary="删除营业审批申请",description="删除营业审批申请",operationId="",produces={"application/json"},
      *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
      *   @SWG\Parameter(in="path",name="id",type="string",description="id",required=true),
      *   @SWG\Response(response="default", description="操作成功")
      * )
      */
     public function uncertify(Request $request, $id){
+        if($this->isIllegal($request->user(),$id)){
+            return $this->sendErrorMsg('这不是您的商店！非法操作！');
+        }
         $shop = Shop::find($id);
         $shop->certify = '未审核';
         $shop->editable = true;
@@ -253,99 +259,15 @@ class ShopController extends Controller
             ]);
     }
 
-    /*
-    * 增加商品*/
-    /**
-     * @SWG\Post(path="/api/shop/{shop_id}/goods", tags={"Shop"},summary="创建我的商店",description="创建商店",operationId="",produces={"application/json"},
-     *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
-     *   @SWG\Parameter(in="path",name="id",type="number",description="商店id",required=true),
-     *   @SWG\Parameter(in="formData",name="name",type="string",description="商品名称",required=true),
-     *   @SWG\Parameter(in="formData",name="main_pic",type="string",description="商品主图",required=true),
-     *   @SWG\Parameter(in="formData",name="big_pic",type="string",description="商品大图",required=false),
-     *   @SWG\Parameter(in="formData",name="description",type="string",description="商品描述",required=false),
-     *   @SWG\Parameter(in="formData",name="content",type="string",description="商品详情",required=false),
-     *   @SWG\Parameter(in="formData",name="status",type="string",description="商品状态",required=true, default="正常"),
-     *   @SWG\Parameter(in="formData",name="type",type="string",description="商品类型",required=true),
-     *   @SWG\Parameter(in="formData",name="tag",type="string",description="商品标签",required=false),
-     *   @SWG\Parameter(in="formData",name="price",type="string",description="商品价格",required=true),
-     *   @SWG\Parameter(in="formData",name="stock",type="string",description="商品库存",required=true),
-     *   @SWG\Parameter(in="formData",name="discount_price",type="string",description="商品折扣价格",required=false),
-     *   @SWG\Response(response="default", description="操作成功")
-     * )
-     */
-    public function createGoods(Request $request, $shopId){
+    public function isIllegal($user, $shopId) {
         $shop = Shop::find($shopId);
-        $data = $request->all();
-        $validator = $this->goodsValidator($data);
-        if($validator->fails()){
-            return $this->sendErrorMsg($validator->errors());
+        if(!$shop){return true; }
+        if($shop->user_id === $user->id){
+            return false;
         }
-        try{
-            $goods = new Goods($data);
-            $goods->save();
-            return $this->sendSuccessMsg('创建成功');
-        }catch(\Exception $e){
-            return $this->sendSqlErrorMsg($e);
+        else{
+            return true;
         }
     }
 
-    /*
-    * 增加商品*/
-    /**
-     * @SWG\Delete(path="/api/shop/{shop_id}/goods/{goods_id}", tags={"Shop"},summary="创建我的商店",description="创建商店",operationId="",produces={"application/json"},
-     *   @SWG\Parameter(in="header",name="Authorization",type="string",description="Token",required=true),
-     *   @SWG\Parameter(in="path",name="shop_id",type="number",description="商店id",required=true),
-     *   @SWG\Parameter(in="path",name="goods_id",type="number",description="商店id",required=true),
-     *   @SWG\Response(response="default", description="操作成功")
-     * )
-     */
-    public function deleteGoods(Request $request, $shopId, $goodsId){
-        $shop = Shop::find($shopId);
-        if(!$shop){return $this->sendErrorMsg('商店不存在！');}
-        $deleteGood = $shop->goods->where('id',$goodsId)->first();
-        if($deleteGood){
-            try{
-                $deleteGood->delete();
-                return $this->sendSuccessMsg('删除成功!');
-            }catch(\Exception $e){
-                return $this->sendSqlErrorMsg($e);
-            }
-        }else{
-            return $this->sendErrorMsg('商品不存在！');
-        }
-    }
-
-    public function goodsValidator(array $data){
-        return Validator::make(
-            $data,
-            [
-                'shop_id'=> 'required',
-                'name' => 'required',
-                'main_pic' => 'required',
-                'big_pic' => 'nullable',
-                'description'=>'nullable',
-                'content'=>'nullable',
-                'status'=>'required',
-                'type'=>'required',
-                'tag'=>'nullable',
-                'price'=>'required|numeric|min:0.01|max:99999999.00',
-                'stock'=>'required|numeric|min:1|max:99999999',
-                'discount_price'=>'nullable',
-            ],
-            [
-                "shop_id.required"=>'商店id不能为空！',
-                "name.required"=>'商品名不能为空！',
-                'main_pic.required'=>'商品主图不能为空！',
-                'status.required'=> '商品状态不能为空！',
-                'type'=> '商品类型不能为空！',
-                'price.required'=> '商品价格不能为空！',
-                'price.numeric'=> '商品价格不能为非数值！',
-                'price.min'=> '商品价格不能为0！',
-                'price.max'=> '商品价格不能为大于99999999！',
-                'stock.numeric'=> '商品库存不能为非整数值类型的值！',
-                'stock.required'=> '商品库存不能为空！',
-                'stock.min'=> '商品库存不能小于1！',
-                'stock.max'=> '商品库存不能超过99999999！'
-            ]);
-    }
 }
